Zipzy-backend--p/requirements.txt
fastapi
uvicorn
pydantic>=2.0.0
python-dotenv
google-generativeai
geopy
numpy
pandas
scikit-learn
python-dateutil
python-multipart
